src/pybind/mgr/dashboard_v2/requirements.txt
astroid==1.6.1
attrs==17.4.0
backports.functools-lru-cache==1.4
cheroot==6.0.0
CherryPy==13.1.0
configparser==3.5.0
coverage==4.4.2
enum34==1.1.6
funcsigs==1.0.2
isort==4.2.15
lazy-object-proxy==1.3.1
mccabe==0.6.1
mock==2.0.0
more-itertools==4.1.0
pbr==3.1.1
pluggy==0.6.0
portend==2.2
py==1.5.2
pycodestyle==2.3.1
pycparser==2.18
pylint==1.8.2
pytest==3.3.2
pytest-cov==2.5.1
python-bcrypt==0.3.2
pytz==2017.3
pyopenssl==16.2.0
requests==2.18.4
singledispatch==3.4.0.3
six==1.11.0
tempora==1.10
tox==2.9.1
virtualenv==15.1.0
wrapt==1.10.11
